Move GitHub databases files to separate directory
