+++ extensions/ql-vscode/src/databases/github-databases/index.ts
@@ -1,0 +1,1 @@
+export * from "./github-database-module";
